refactor: highlight 타입, 유틸 함수 분리
