+++ frontend/src/types/highlight.ts
@@ -1,0 +1,9 @@
+export interface Highlight {
+  start: number;
+  length: number;
+}
+
+export interface EditorBlock {
+  text: string;
+  highlightList: Highlight[];
+}
