+++ frontend/src/types/index.ts
@@ -4,4 +4,4 @@ export * from './emotion';
 export * from './styles';
 export * from './essentialPropsWithChildren';
 export * from './reviewGroup';
-export * from './urlGeneratorForm';
+export * from './highlight';
